requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
prompt-toolkit>=3.0.0
networkx>=3.0

# Optional: for semantic search (GCC codebase manager)
# sentence-transformers>=2.0.0
